yolo_cropDisease_detection_web/yolo_cropDisease_detection_web/yolo_cropDisease_detection_flask/requirements.txt
flask
flask-cors
flask-socketio
ultralytics
opencv-python
numpy
Pillow
requests
onnxruntime
zai
keyring
